requirements.txt
djangorestframework==3.15.1
django-cors-headers==4.3.1
mysql-connector-python==8.3.0
mysqlclient==2.2.4
djangorestframework_simplejwt==5.3.1
cryptography==42.0.5
mailslurp-client==15.19.4
django-ratelimit==4.1.0
pytz==2024.1